requirements.txt
Flask
requests
azure-kusto-data